Lots of bug fixes, now appends rather than creating new awds

+++ ethoHelper.py
@@ -16,13 +16,17 @@ def converter(inputFile, outputFile, startTime, date, awdFolder, buffer):
     animDict[animNum] = []
 
     print("Creating internal data storage...")
-    while worksheet.cell(row=(x + 1), column=1).value == "Result 1":
+    correctCol = 5
+    for col in range(5, 10):
+        if (isinstance(worksheet.cell(x + 1, col).value, float) or isinstance(worksheet.cell(x + 1, col).value, int)):
+            correctCol = col
+    while worksheet.cell(row=(x+1),column=1).value != None:
         if animNum == worksheet.cell(row = x + 1,column = 3).value:
-            animDict[animNum].append(worksheet.cell(row=x + 1, column=6).value)
+            animDict[animNum].append(worksheet.cell(row=x + 1, column=correctCol).value)
         else:
             animNum = worksheet.cell(row = x + 1, column = 3).value
             animDict[animNum] = []
-            animDict[animNum].append(worksheet.cell(row=x + 1,column=6).value)
+            animDict[animNum].append(worksheet.cell(row=x + 1,column=correctCol).value)
         x += 1
 
     animNumSheet = 2
@@ -37,7 +41,7 @@ def converter(inputFile, outputFile, startTime, date, awdFolder, buffer):
     print("Appending to checktime file...")
     for lst in animList:
         i = startI
-        for count in range(len(animDict) + 1):
+        for count in range(len(animDict) + 2):
             if newSheet.cell(row=1, column=count + 1).value == lst:
                 currCount = count
         for gapCount in range(int(buffer)):
@@ -58,20 +62,20 @@ def converter(inputFile, outputFile, startTime, date, awdFolder, buffer):
         for lst in animList:
             path = awdFolder + "\\" + lst + ".awd"
             print("Generating awd file: " + path)
-            currFile = open(path, 'w')
-            currFile.write(lst + '\n')
-            currFile.write(date + '\n')
-            currFile.write(startTime + '\n')
-            currFile.write("4\n")
-            currFile.write("0\n")
-            currFile.write("Ignore\n")
-            currFile.write("M\n")
-            for zeroC in range(6):
+            if not os.path.exists(path):
+                currFile = open(path, 'w')
+                currFile.write(lst + '\n')
+                currFile.write(date + '\n')
+                currFile.write(startTime + '\n')
+                currFile.write("4\n")
                 currFile.write("0\n")
+                currFile.write("Ignore\n")
+                currFile.write("M\n")
+            else:
+                currFile = open(path, 'a')
             for data in animDict[lst]:
                 currFile.write(str(data) + '\n')
             currFile.close()
-
 def test():
     print("Test!")
     converter("ex.xlsx", "Checktime.xlsx", "19:00", "05/18/2018", "AWDFiles", 3)
@@ -91,7 +95,7 @@ def main():
         startTime = input("Please enter start time of first day(ex: 09:16,23:14): ")
     date = input("Please enter start date (Leave blank for no AWD file creation)(ex: 01/02/1970): ")
     if date != "":
-        print("AWDs will be generated in format: <folder name>/<column name>.awd")
+        print("AWDs will be found and generated in format: <folder name>/<column name>.awd")
         awdFolder = input("Please enter folder name for AWD file storage(default: AWDFiles): ")
         if awdFolder == "":
             awdFolder = "AWDFiles"
@@ -100,4 +104,4 @@ def main():
         awdFolder = ""
     converter(inputFile + ".xlsx", outputFile + ".xlsx", startTime, date, awdFolder, buffer)
 
-main()
+test()
